review(#932): split flow_evaluator.rs into module (Nico #3)
2721-line file → 6-file module. Public surface preserved for external
consumers (run_engine_proposal_pass, LlmProposalHint, SatisfiedTransition);
internal helpers (build_query_input_for_requires, cardinality_satisfied,
evidence_hash, RequiresQueryable, evaluate_* variants, write_engine_proposal)
no longer re-exported since no crate code outside this module references them.

Layout:
- primitives.rs (223) — SatisfiedTransition, build_query_input_for_requires,
  cardinality_satisfied, evidence_hash — pure fns + types, no async
- queryable.rs (224) — RequiresQueryable trait + evaluate_* async layer
- engine_pass.rs (338) — run_engine_proposal_pass + write_engine_proposal
  + LlmProposalHint (the composed load→evaluate→write pipeline)
- unit_tests.rs (853) — trait-stubbed unit coverage
- e2e_tests.rs (1051) — real-PerspectiveInstance coverage
- mod.rs (78) — module doc + re-exports

cargo check --features generate_snapshot: clean (0 errors)
cargo fmt --all --check: clean
